+++ finagle-core/src/main/scala/com/twitter/finagle/server/ListeningStackServer.scala
@@ -3,15 +3,18 @@ package com.twitter.finagle.server
 import com.twitter.finagle.Stack.Param
 import com.twitter.finagle.filter.RequestLogger
 import com.twitter.finagle.param._
-import com.twitter.finagle.{ClientConnection, ListeningServer, ServiceFactory, Stack}
+import com.twitter.finagle.ClientConnection
+import com.twitter.finagle.ListeningServer
+import com.twitter.finagle.ServiceFactory
+import com.twitter.finagle.Stack
 import com.twitter.finagle.stack.Endpoint
-import com.twitter.finagle.stats.{
-  RelativeNameMarkingStatsReceiver,
-  RoleConfiguredStatsReceiver,
-  Server
-}
+import com.twitter.finagle.stats.RelativeNameMarkingStatsReceiver
+import com.twitter.finagle.stats.RoleConfiguredStatsReceiver
+import com.twitter.finagle.stats.Server
 import com.twitter.util.registry.GlobalRegistry
-import com.twitter.util.{CloseAwaitably, Future, Time}
+import com.twitter.util.CloseAwaitably
+import com.twitter.util.Future
+import com.twitter.util.Time
 import java.net.SocketAddress
 
 /**
@@ -65,10 +68,14 @@ trait ListeningStackServer[Req, Rep, This <: ListeningStackServer[Req, Rep, This
             Server,
             Some(serverLabel))
 
-      private[this] val serverParams = params +
-        Label(serverLabel) +
-        Stats(statsReceiver) +
-        Monitor(reporter(label, None).andThen(monitor))
+      private[this] val serverParams = StackServer.DefaultInjectors.injectors.foldLeft(
+        params +
+          Label(serverLabel) +
+          Stats(statsReceiver) +
+          Monitor(reporter(label, None).andThen(monitor))) {
+        case (prms, injector) =>
+          injector(prms)
+      }
 
       // We re-parameterize in case `newListeningServer` needs to access the
       // finalized parameters.