+++ finagle-core/src/main/scala/com/twitter/finagle/server/StackServer.scala
@@ -14,6 +14,7 @@ import com.twitter.finagle.tracing._
 import com.twitter.finagle.Stack
 import com.twitter.finagle._
 import com.twitter.jvm.Jvm
+import scala.collection.immutable
 
 object StackServer {
   private[this] lazy val newJvmFilter = new MkJvmFilter(Jvm())
@@ -187,6 +188,19 @@ object StackServer {
    * A set of StackTransformers for transforming server stacks.
    */
   private[finagle] object DefaultTransformer extends StackTransformerCollection
+
+  /**
+   * A set of ServerParamsInjectors for transforming server params.
+   */
+  private[finagle] object DefaultInjectors {
+    @volatile private var underlying = immutable.Queue.empty[ServerParamsInjector]
+
+    def append(injector: ServerParamsInjector): Unit =
+      synchronized { underlying = underlying :+ injector }
+
+    def injectors: Seq[ServerParamsInjector] =
+      underlying
+  }
 }
 
 /**